requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
beautifulsoup4==4.13.4
    # via -r requirements.in
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.2
    # via requests
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.177.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.3
    # via playwright
grpcio==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
imageio==2.37.0
    # via scikit-image
iniconfig==2.1.0
    # via pytest
lazy-loader==0.4
    # via scikit-image
networkx==3.5
    # via scikit-image
numpy==2.2.6
    # via
    #   imageio
    #   opencv-python
    #   scikit-image
    #   scipy
    #   tifffile
opencv-python==4.12.0.88
    # via -r requirements.in
packaging==25.0
    # via
    #   lazy-loader
    #   pytest
    #   scikit-image
pillow==11.3.0
    # via
    #   imageio
    #   scikit-image
playwright==1.54.0
    # via
    #   -r requirements.in
    #   pytest-playwright
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via google-generativeai
pydantic-core==2.33.2
    # via pydantic
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via pytest
pyparsing==3.2.3
    # via httplib2
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-base-url
    #   pytest-playwright
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-playwright==0.7.0
    # via -r requirements.in
python-dotenv==1.1.1
    # via -r requirements.in
python-slugify==8.0.4
    # via pytest-playwright
requests==2.32.4
    # via
    #   google-api-core
    #   pytest-base-url
rsa==4.9.1
    # via google-auth
scikit-image==0.25.2
    # via -r requirements.in
scipy==1.16.1
    # via scikit-image
soupsieve==2.7
    # via beautifulsoup4
text-unidecode==1.3
    # via python-slugify
tifffile==2025.6.11
    # via scikit-image
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.14.1
    # via
    #   beautifulsoup4
    #   google-generativeai
    #   pydantic
    #   pydantic-core
    #   pyee
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
